fix(work): file links open in a new tab from a2ui Markdown too
Workspace file paths inside the a2ui Markdown component (Hermes
synthetic-markdown surface case) rendered as plain anchors without the
target="_blank" + popup-blocker fallback wiring — clicks navigated
the chat tab away to the PDF instead of opening it alongside.

Move the components map (and openFileLink) into
apps/web/src/lib/linkify-workspace-paths.tsx so it can be imported
from both render sites, and apply it to the a2ui ReactMarkdown
invocation. Now every clickable workspace path opens in a new tab
regardless of which renderer produced the markdown.

Renamed the helper file from .ts to .tsx since it now exports JSX.
