+++ apps/web/src/a2ui/components.tsx
@@ -10,7 +10,10 @@
 
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
-import { linkifyWorkspacePaths } from "@/lib/linkify-workspace-paths";
+import {
+  WORKSPACE_MARKDOWN_COMPONENTS,
+  linkifyWorkspacePaths,
+} from "@/lib/linkify-workspace-paths";
 import { registerComponent, renderChildren } from "./renderer";
 import type { RenderContext } from "./renderer";
 import type { A2UIComponent } from "./types";
@@ -38,7 +41,10 @@ registerComponent("Markdown", (comp: A2UIComponent) => {
   const props = comp as unknown as MarkdownProps & { id: string };
   return (
     <div key={props.id} className="work-markdown">
-      <ReactMarkdown remarkPlugins={[remarkGfm]}>
+      <ReactMarkdown
+        remarkPlugins={[remarkGfm]}
+        components={WORKSPACE_MARKDOWN_COMPONENTS}
+      >
         {linkifyWorkspacePaths(props.text)}
       </ReactMarkdown>
     </div>
